[Pal/Linux-SGX] Sanitize return values of ocalls
When doing an ocall, untrusted part could return an invalid value for
example `write` could return size bigger than passed to it. This commit
adds checks for such invalid values.

+++ Pal/src/host/Linux-SGX/ocall_types.h
@@ -252,7 +252,7 @@ typedef struct {
 
 typedef struct {
     struct pollfd* ms_fds;
-    int ms_nfds;
+    size_t ms_nfds;
     int64_t ms_timeout_us;
 } ms_ocall_poll_t;
